Split linkers out from compilers.py
